fix(review): move step extraction to artifact module
