+++ src/autoskillit/server/_recipe_artifact.py
@@ -26,7 +26,9 @@ from autoskillit.core import (
     RecipeFlowGeneration,
     atomic_write,
     build_recipe_execution_credential,
+    fast_dumps,
     get_logger,
+    load_yaml,
     validate_recipe_artifact_sections,
 )
 from autoskillit.server._recipe_generation import (
@@ -47,6 +49,74 @@ class RecipeArtifactError(RuntimeError):
 
 class RecipeArtifactSchemaError(RecipeArtifactError):
     """A recipe artifact violates the static pullable-section schema."""
+
+
+class RecipeStepExtractionError(Exception):
+    """A persisted recipe body could not be parsed or serialized."""
+
+    def __init__(self, code: str, detail: str) -> None:
+        super().__init__(detail)
+        self.code = code
+
+
+def _persisted_recipe_steps(persisted: dict[str, Any]) -> dict[str, Any]:
+    content = persisted.get("content", "") or ""
+    if not isinstance(content, str) or not content:
+        return {}
+    try:
+        parsed = load_yaml(content)
+    except Exception as exc:
+        get_logger(__name__).warning("recipe_step_yaml_parse_failed", exc_info=True)
+        raise RecipeStepExtractionError(
+            "recipe_artifact_parse_failed", f"{type(exc).__name__}: {exc}"
+        ) from exc
+    if not isinstance(parsed, dict):
+        raise RecipeStepExtractionError(
+            "recipe_artifact_parse_failed", "recipe content is not a mapping"
+        )
+    steps = parsed.get("steps")
+    if not isinstance(steps, dict):
+        raise RecipeStepExtractionError(
+            "recipe_artifact_parse_failed", "recipe steps are not a mapping"
+        )
+    return steps
+
+
+def extract_recipe_step_bodies(
+    persisted: dict[str, Any],
+    ordered_step_names: tuple[str, ...],
+) -> tuple[tuple[str, str], ...]:
+    """Extract canonical compact bodies for ordered persisted steps."""
+    steps = _persisted_recipe_steps(persisted)
+    bodies: list[tuple[str, str]] = []
+    for step_name in ordered_step_names:
+        step_obj = steps.get(step_name)
+        if step_obj is None:
+            continue
+        if not isinstance(step_obj, dict):
+            raise RecipeStepExtractionError(
+                "recipe_section_serialization_failed", "recipe step is not a mapping"
+            )
+        try:
+            bodies.append((step_name, fast_dumps({step_name: step_obj})))
+        except Exception as exc:
+            get_logger(__name__).warning(
+                "recipe_step_yaml_serialize_failed",
+                step_name=step_name,
+                exc_info=True,
+            )
+            raise RecipeStepExtractionError(
+                "recipe_section_serialization_failed", f"{type(exc).__name__}: {exc}"
+            ) from exc
+    return tuple(bodies)
+
+
+def extract_step_body_from_persisted(persisted: dict[str, Any], step_name: str) -> str:
+    """Extract one compact step body from a persisted recipe artifact."""
+    if not step_name:
+        return ""
+    bodies = extract_recipe_step_bodies(persisted, (step_name,))
+    return bodies[0][1] if bodies else ""
 
 
 @dataclass(frozen=True, slots=True)
@@ -636,8 +706,11 @@ __all__ = [
     "PreparedRecipeGeneration",
     "RecipeArtifactError",
     "RecipeArtifactSchemaError",
+    "RecipeStepExtractionError",
     "build_canonical_recipe_artifact_payload",
     "build_recipe_flow_generation",
+    "extract_recipe_step_bodies",
+    "extract_step_body_from_persisted",
     "load_recipe_artifact",
     "persist_recipe_artifact",
     "prepare_recipe_delivery_generation",
